requirements.txt
Django==3.2
django-extensions==3.1.5
ipython==7.31.0
django-pagedown==2.2.1
django-markdown-deux==1.0.5